pyproject.toml
[tool.poetry]
name = "GameInfo"
version = "1.0.7"
description = "System info tool for gamers"
authors = ["Michael John <michael.john@gmx.at>"]
license = "GPL"
include = ["images/*", "gameinfo/locales/de/LC_MESSAGES/*.mo"]

[tool.poetry.scripts]
GameInfo = 'gameinfo.__main__:main'

[tool.poetry.dependencies]
python = "^3.7"
python-gettext = "^4.0"
requests = "^2.27.1"
ttkbootstrap = "^1.7.6"
CairoSVG = "^2.5.2"
Pillow = "^9.1.0"
cffi = "^1.15.0"
vdf = "^3.4"
importlib-metadata = "^4.11.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-dynamic-versioning]
enable = false

[tool.poetry.urls]
homepage = "https://github.com/amstelchen/GameInfo"
repository = "https://github.com/amstelchen/GameInfo"
documentation = "https://github.com/amstelchen/GameInfo"
